// enums/exercise-category.ts
export enum ExerciseCategory {
    Chest = 'Chest',
    Back = 'Back',
    Shoulders = 'Shoulders',
    Biceps = 'Biceps',
    Triceps = 'Triceps',
    Quadriceps = 'Quadriceps',
    Hamstrings = 'Hamstrings',
    Calves = 'Calves',
    Glutes = 'Glutes',
    Abs = 'Abs',
    Cardio = 'Cardio',
    Miscellaneous = 'Miscellaneous',
}